backend/pyproject.toml
[project]
name = "resource-wise-backend"
description = "FastAPI backend for Resource Wise"
authors = [
    { name = "Jatin Gandhi" }
]
version = "0.1.0"
requires-python = "==3.13.*"

dependencies = [
    # Core FastAPI
    "fastapi>=0.115.12",
    "uvicorn>=0.34.3",
    # Database & ORM
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9", # Required for Alembic migrations
    # Vector Database (for AI embeddings)
    "pgvector>=0.3.6",
    # AI & Embeddings
    "openai>=1.52.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.15",
    # HTTP Client (for OpenAI API calls)
    "httpx>=0.27.0",
    # Configuration & Environment
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.6.1",
    # Logging
    "structlog>=24.4.0",
    "greenlet>=3.2.3",
]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    ".venv",
    "venv", 
    "__pycache__",
    "alembic/versions/*",
    "alembic/versions/**/*.py",
]

ignore = ["B008", "B904"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings  
    "F",  # Pyflakes
    "I",  # isort (import sorting)
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]

[tool.ruff.lint.isort]
force-wrap-aliases = true
combine-as-imports = true
